pyproject.toml
[project]
name = "recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "tqdm>=4.67.3",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "numpy>=2.4.2",
    "pandas>=3.0.0",
    "SQLAlchemy==2.0.49",
    "sqlparse==0.5.5",
    "python-dotenv==1.2.2",
    "datasets>=4.8.5",
    "tdqm>=0.0.1",
    "supabase>=2.30.0",
]
